+++ R/utils.R
@@ -1,4 +1,4 @@
-# Make a sparse_model_matrix for fixest estimate
+# Make a sparse_model_matrix for fixest estimate. This only keeps the variables that are not removed from `fixest::feols`
 sparse_model_matrix = function(data, fixest) {
 	Z = NULL
 
@@ -11,7 +11,7 @@ sparse_model_matrix = function(data, fixest) {
 
 		Z_fixef = Matrix::sparse.model.matrix(frmla, data = data)
 
-		temp = fixest::fixef(first_stage_est)
+		temp = fixest::fixef(fixest)
 		select =	lapply(names(temp), function(var){
 			names = names(temp[[var]])
 			names = names[temp[[var]] != 0]
@@ -20,7 +20,7 @@ sparse_model_matrix = function(data, fixest) {
 		})
 
 
-		Z = cbind(Z, Z_fixef[, unlist(blank)])
+		Z = cbind(Z, Z_fixef[, unlist(select)])
 	}
 
 	return(Z)