+++ data-raw/gen_data.R
@@ -1,7 +1,10 @@
 # Homogeneous Effects ----------------------------------------------------------
 
+set.seed(1)
 df_hom <- gen_data(panel = c(1990, 2020), g1 = 2000, g2 = 2010, g3 = 0, te1 = 2, te2 = 2, te3 = 0, te_m1 = 0, te_m2 = 0, te_m3 = 0)
 
+# Fake weights
+df_hom <- df_hom %>% dplyr::group_by(unit) %>% dplyr::mutate(weight = runif(1))
 
 # Heterogeneous Effects --------------------------------------------------------